content/testsuite: include small blob test in standard suite
